+++ emencia/django/countries/admin.py
@@ -2,13 +2,16 @@
 from django.contrib import admin
 from django.utils.translation import ugettext as _
 
+from parler.admin import TranslatableAdmin
+
 from emencia.django.countries.models import Country
 
 
-class CountryAdmin(admin.ModelAdmin):
-    search_fields = ('printable_name', 'name', 'iso', 'iso3')
+class CountryAdmin(TranslatableAdmin):
     list_display = ('printable_name', 'name', 'iso', 'iso3', 'numcode','level')
-    ordering = ('printable_name',)
+    search_fields = (
+        'translations__printable_name', 'translations__name', 'iso', 'iso3')
+
     fieldsets = ((None, {'fields': ('printable_name', 'name')}),
                  (_('References'), {'fields': ('iso', 'iso3', 'numcode')}),
                  (None, {'fields': ('level',)}),)